Make it possible to specify your own naming convention for queues, topics  and subscriptions. This will allow Nimbus to be used in environments that have existing implementations.
